# generated-model-resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Denaly | API",
    "description": "Swagger UI for Denaly API",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "API Support",
      "url": "https://www.example.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:4000/web",
      "description": "Local server"
    },
    {
      "url": "https://development.gigantic-server.com",
      "description": "Development server"
    },
    {
      "url": "https://staging.gigantic-server.com",
      "description": "Staging server"
    },
    {
      "url": "https://api.gigantic-server.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all Users",
        "description": "Get all Users",
        "operationId": "getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of Users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserEntity"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new User",
        "description": "Create a new User",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "stripe_customer_account_id": "",
                "stripe_account_id": "",
                "stripe_account_verified": true,
                "stripe_identity_verified": true,
                "first_name": "",
                "last_name": "",
                "bio": "",
                "icon_s3object_id": 1,
                "town": "",
                "city": "",
                "state": "",
                "zipcode": 1,
                "country": "",
                "tags": "",
                "specialties": "",
                "person_verified": true,
                "email_verified": true,
                "phone_verified": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get User by ID",
        "description": "Get User by ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update User by ID",
        "description": "Update User by ID",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "stripe_customer_account_id": "",
                "stripe_account_id": "",
                "stripe_account_verified": true,
                "stripe_identity_verified": true,
                "first_name": "",
                "last_name": "",
                "bio": "",
                "icon_s3object_id": 1,
                "town": "",
                "city": "",
                "state": "",
                "zipcode": 1,
                "country": "",
                "tags": "",
                "specialties": "",
                "person_verified": true,
                "email_verified": true,
                "phone_verified": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Patch User by ID",
        "description": "Patch User by ID",
        "operationId": "PatchUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "stripe_customer_account_id": "",
                "stripe_account_id": "",
                "stripe_account_verified": true,
                "stripe_identity_verified": true,
                "first_name": "",
                "last_name": "",
                "bio": "",
                "icon_s3object_id": 1,
                "town": "",
                "city": "",
                "state": "",
                "zipcode": 1,
                "country": "",
                "tags": "",
                "specialties": "",
                "person_verified": true,
                "email_verified": true,
                "phone_verified": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete User by ID",
        "description": "Delete User by ID",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/user-auth-providers": {
      "get": {
        "tags": [
          "UserAuthProvider"
        ],
        "summary": "Get all UserAuthProviders",
        "description": "Get all UserAuthProviders",
        "operationId": "getUserAuthProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of UserAuthProviders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAuthProviderEntity"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UserAuthProvider"
        ],
        "summary": "Create a new UserAuthProvider",
        "description": "Create a new UserAuthProvider",
        "operationId": "createUserAuthProvider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserAuthProviderDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "details": "",
                "user_id": 1,
                "provider_name": "",
                "provider_id": ""
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created UserAuthProvider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuthProviderEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/user-auth-providers/{id}": {
      "get": {
        "tags": [
          "UserAuthProvider"
        ],
        "summary": "Get UserAuthProvider by ID",
        "description": "Get UserAuthProvider by ID",
        "operationId": "getUserAuthProviderById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UserAuthProvider found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuthProviderEntity"
                }
              }
            }
          },
          "404": {
            "description": "Userauthprovider not found"
          }
        }
      },
      "put": {
        "tags": [
          "UserAuthProvider"
        ],
        "summary": "Update UserAuthProvider by ID",
        "description": "Update UserAuthProvider by ID",
        "operationId": "updateUserAuthProvider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserAuthProviderDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "details": "",
                "user_id": 1,
                "provider_name": "",
                "provider_id": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "UserAuthProvider updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuthProviderEntity"
                }
              }
            }
          },
          "404": {
            "description": "UserAuthProvider not found"
          }
        }
      },
      "patch": {
        "tags": [
          "UserAuthProvider"
        ],
        "summary": "Patch UserAuthProvider by ID",
        "description": "Patch UserAuthProvider by ID",
        "operationId": "PatchUserAuthProvider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserAuthProviderDto"
              },
              "example": {
                "id": 1,
                "metadata": "",
                "create_at": "",
                "updated_at": "",
                "deleted_at": "",
                "details": "",
                "user_id": 1,
                "provider_name": "",
                "provider_id": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "UserAuthProvider updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuthProviderEntity"
                }
              }
            }
          },
          "404": {
            "description": "UserAuthProvider not found"
          }
        }
      },
      "delete": {
        "tags": [
          "UserAuthProvider"
        ],
        "summary": "Delete UserAuthProvider by ID",
        "description": "Delete UserAuthProvider by ID",
        "operationId": "deleteUserAuthProvider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of {model_name}",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "UserAuthProvider deleted"
          },
          "404": {
            "description": "UserAuthProvider not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserEntity": {
        "type": "object",
        "description": "User entity",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "stripe_customer_account_id": {
            "type": "string",
            "format": "string",
            "description": "Stripe customer account id"
          },
          "stripe_account_id": {
            "type": "string",
            "format": "string",
            "description": "Stripe account id"
          },
          "stripe_account_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Stripe account verified"
          },
          "stripe_identity_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Stripe identity verified"
          },
          "first_name": {
            "type": "string",
            "format": "string",
            "description": "First name"
          },
          "last_name": {
            "type": "string",
            "format": "string",
            "description": "Last name"
          },
          "bio": {
            "type": "string",
            "format": "string",
            "description": "Bio"
          },
          "icon_s3object_id": {
            "type": "integer",
            "format": "int64",
            "description": "Icon s3object id"
          },
          "town": {
            "type": "string",
            "format": "string",
            "description": "Town"
          },
          "city": {
            "type": "string",
            "format": "string",
            "description": "City"
          },
          "state": {
            "type": "string",
            "format": "string",
            "description": "State"
          },
          "zipcode": {
            "type": "integer",
            "format": "int64",
            "description": "Zipcode"
          },
          "country": {
            "type": "string",
            "format": "string",
            "description": "Country"
          },
          "tags": {
            "type": "string",
            "format": "string",
            "description": "Tags"
          },
          "specialties": {
            "type": "string",
            "format": "string",
            "description": "Specialties"
          },
          "person_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Person verified"
          },
          "email_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Email verified"
          },
          "phone_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Phone verified"
          }
        },
        "required": [
          "id",
          "create_at",
          "stripe_account_verified",
          "stripe_identity_verified",
          "person_verified",
          "email_verified",
          "phone_verified"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "description": "Create User DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "stripe_customer_account_id": {
            "type": "string",
            "format": "string",
            "description": "Stripe customer account id"
          },
          "stripe_account_id": {
            "type": "string",
            "format": "string",
            "description": "Stripe account id"
          },
          "stripe_account_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Stripe account verified"
          },
          "stripe_identity_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Stripe identity verified"
          },
          "first_name": {
            "type": "string",
            "format": "string",
            "description": "First name"
          },
          "last_name": {
            "type": "string",
            "format": "string",
            "description": "Last name"
          },
          "bio": {
            "type": "string",
            "format": "string",
            "description": "Bio"
          },
          "icon_s3object_id": {
            "type": "integer",
            "format": "int64",
            "description": "Icon s3object id"
          },
          "town": {
            "type": "string",
            "format": "string",
            "description": "Town"
          },
          "city": {
            "type": "string",
            "format": "string",
            "description": "City"
          },
          "state": {
            "type": "string",
            "format": "string",
            "description": "State"
          },
          "zipcode": {
            "type": "integer",
            "format": "int64",
            "description": "Zipcode"
          },
          "country": {
            "type": "string",
            "format": "string",
            "description": "Country"
          },
          "tags": {
            "type": "string",
            "format": "string",
            "description": "Tags"
          },
          "specialties": {
            "type": "string",
            "format": "string",
            "description": "Specialties"
          },
          "person_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Person verified"
          },
          "email_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Email verified"
          },
          "phone_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Phone verified"
          }
        }
      },
      "UpdateUserDto": {
        "type": "object",
        "description": "Update User DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "stripe_customer_account_id": {
            "type": "string",
            "format": "string",
            "description": "Stripe customer account id"
          },
          "stripe_account_id": {
            "type": "string",
            "format": "string",
            "description": "Stripe account id"
          },
          "stripe_account_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Stripe account verified"
          },
          "stripe_identity_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Stripe identity verified"
          },
          "first_name": {
            "type": "string",
            "format": "string",
            "description": "First name"
          },
          "last_name": {
            "type": "string",
            "format": "string",
            "description": "Last name"
          },
          "bio": {
            "type": "string",
            "format": "string",
            "description": "Bio"
          },
          "icon_s3object_id": {
            "type": "integer",
            "format": "int64",
            "description": "Icon s3object id"
          },
          "town": {
            "type": "string",
            "format": "string",
            "description": "Town"
          },
          "city": {
            "type": "string",
            "format": "string",
            "description": "City"
          },
          "state": {
            "type": "string",
            "format": "string",
            "description": "State"
          },
          "zipcode": {
            "type": "integer",
            "format": "int64",
            "description": "Zipcode"
          },
          "country": {
            "type": "string",
            "format": "string",
            "description": "Country"
          },
          "tags": {
            "type": "string",
            "format": "string",
            "description": "Tags"
          },
          "specialties": {
            "type": "string",
            "format": "string",
            "description": "Specialties"
          },
          "person_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Person verified"
          },
          "email_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Email verified"
          },
          "phone_verified": {
            "type": "boolean",
            "format": "boolean",
            "description": "Phone verified"
          }
        }
      },
...
      "UserAuthProviderEntity": {
        "type": "object",
        "description": "UserAuthProvider entity",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "details": {
            "type": "string",
            "format": "string",
            "description": "Details"
          },
          "user_id": {
            "type": "integer",
            "format": "int64",
            "description": "User id"
          },
          "provider_name": {
            "type": "string",
            "format": "string",
            "description": "Provider name"
          },
          "provider_id": {
            "type": "string",
            "format": "string",
            "description": "Provider id"
          }
        },
        "required": [
          "id",
          "create_at",
          "user_id",
          "provider_name",
          "provider_id"
        ]
      },
      "CreateUserAuthProviderDto": {
        "type": "object",
        "description": "Create UserAuthProvider DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "details": {
            "type": "string",
            "format": "string",
            "description": "Details"
          },
          "user_id": {
            "type": "integer",
            "format": "int64",
            "description": "User id"
          },
          "provider_name": {
            "type": "string",
            "format": "string",
            "description": "Provider name"
          },
          "provider_id": {
            "type": "string",
            "format": "string",
            "description": "Provider id"
          }
        }
      },
      "UpdateUserAuthProviderDto": {
        "type": "object",
        "description": "Update UserAuthProvider DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Id"
          },
          "metadata": {
            "type": "string",
            "format": "string",
            "description": "Metadata"
          },
          "create_at": {
            "type": "string",
            "format": "date-time",
            "description": "Create at"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Updated at"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Deleted at"
          },
          "details": {
            "type": "string",
            "format": "string",
            "description": "Details"
          },
          "user_id": {
            "type": "integer",
            "format": "int64",
            "description": "User id"
          },
          "provider_name": {
            "type": "string",
            "format": "string",
            "description": "Provider name"
          },
          "provider_id": {
            "type": "string",
            "format": "string",
            "description": "Provider id"
          }
        }
      },
...
    }
  }
}